added bed listener to night manager
